Add java-se module Move classes which contain APIs not supported in Android here.
The java-se module hosts classes which are part of Java Standard Edition but not of the JDK used in the Android SDK.
For code re-use in Android we isolate those classes to that module.
