+++ apps/desktop/desktop-app/build.gradle.kts
@@ -35,6 +35,7 @@ javafx {
 
 dependencies {
     implementation("bisq:persistence")
+    implementation("bisq:java-se")
     implementation("bisq:i18n")
     implementation("bisq:security")
     implementation("bisq:identity")
