+++ apps/oracle-node-app/build.gradle.kts
@@ -16,6 +16,7 @@ application {
 
 dependencies {
     implementation("bisq:persistence")
+    implementation("bisq:java-se")
     implementation("bisq:security")
     implementation("bisq:identity")
     implementation("bisq:bonded-roles")
