+++ java-se/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("bisq.java-library")
+}
+
+dependencies {
+    implementation("bisq:common")
+}
