+++ settings.gradle.kts
@@ -32,6 +32,7 @@ include("daemon")
 include("evolution")
 include("identity")
 include("i18n")
+include("java-se")
 include("offer")
 include("persistence")
 include("platform")
